+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2023-05-21
+
 ### Added
 
 - `VectorIterator` and `VectorRefIterator` for iterating over `Vector`